+++ go/api/response_handlers.go
@@ -91,6 +91,8 @@ func parseInterface(response *C.struct_CommandResponse) (interface{}, error) {
 		return parseMap(response)
 	case C.Sets:
 		return parseSet(response)
+	case C.Ok:
+		return "OK", nil
 	}
 
 	return nil, &errors.RequestError{Msg: "Unexpected return type from Valkey"}
@@ -168,6 +170,27 @@ func handleStringResponse(response *C.struct_CommandResponse) (string, error) {
 
 func handleStringOrNilResponse(response *C.struct_CommandResponse) (Result[string], error) {
 	defer C.free_command_response(response)
+
+	return convertCharArrayToString(response, true)
+}
+
+func handleOkResponse(response *C.struct_CommandResponse) (string, error) {
+	defer C.free_command_response(response)
+
+	typeErr := checkResponseType(response, C.Ok, false)
+	if typeErr != nil {
+		return DefaultStringResponse, typeErr
+	}
+
+	return "OK", nil
+}
+
+func handleOkOrStringOrNilResponse(response *C.struct_CommandResponse) (Result[string], error) {
+	defer C.free_command_response(response)
+
+	if response.response_type == uint32(C.Ok) {
+		return CreateStringResult("OK"), nil
+	}
 
 	return convertCharArrayToString(response, true)
 }
@@ -1411,6 +1434,33 @@ func handleRawStringArrayMapResponse(response *C.struct_CommandResponse) (map[st
 		return nil, err
 	}
 	mapResult, ok := result.(map[string][]string)
+	if !ok {
+		return nil, &errors.RequestError{Msg: "Unexpected conversion result type"}
+	}
+
+	return mapResult, nil
+}
+
+func handleMapOfStringMapResponse(response *C.struct_CommandResponse) (map[string]map[string]string, error) {
+	defer C.free_command_response(response)
+	typeErr := checkResponseType(response, C.Map, false)
+	if typeErr != nil {
+		return nil, typeErr
+	}
+
+	data, err := parseMap(response)
+	if err != nil {
+		return nil, err
+	}
+
+	result, err := mapConverter[map[string]string]{
+		next:     mapConverter[string]{},
+		canBeNil: false,
+	}.convert(data)
+	if err != nil {
+		return nil, err
+	}
+	mapResult, ok := result.(map[string]map[string]string)
 	if !ok {
 		return nil, &errors.RequestError{Msg: "Unexpected conversion result type"}
 	}