+++ java/client/src/test/java/glide/managers/ConnectionManagerTest.java
@@ -111,7 +111,6 @@ public class ConnectionManagerTest {
         ConnectionRequest expectedProtobufConnectionRequest =
                 ConnectionRequest.newBuilder()
                         .setTlsMode(TlsMode.NoTls)
-                        .setConnectionTimeout(250)
                         .setClusterModeEnabled(true)
                         .setReadFrom(ConnectionRequestOuterClass.ReadFrom.Primary)
                         .build();
@@ -404,7 +403,6 @@ public class ConnectionManagerTest {
         ConnectionRequest expectedProtobufConnectionRequest =
                 ConnectionRequest.newBuilder()
                         .setTlsMode(TlsMode.InsecureTls)
-                        .setConnectionTimeout(250)
                         .setClusterModeEnabled(true)
                         .setReadFrom(ConnectionRequestOuterClass.ReadFrom.Primary)
                         .build();