+++ python/tests/test_async_client.py
@@ -83,6 +83,7 @@ from glide.routes import (
     SlotKeyRoute,
     SlotType,
 )
+
 from tests.conftest import create_client
 from tests.utils.utils import (
     check_function_list_response,